[fix]: update integration points after restructuring
